+++ src/game/game_1531a0.c
@@ -269,7 +269,7 @@ void fontLoad(u8 *romstart, u8 *romend, struct font **font1, struct font2a4 **fo
 	struct font *font = malloc(len, MEMPOOL_STAGE);
 	struct font2a4 *font2a4 = font->unk2a4;
 
-	dmaExec(font, romstart, len);
+	dmaExec(font, (u32) romstart, len);
 
 	// Convert pointers
 	for (i = 0; i < 94; i++) {
